+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added RxNorm Map family of functions  
 * Introduced RxNorm processing built on the `PL/R` 
 extension.  
+* Added postprocessing steps are part of `run_setup()`  
 * Fixed bug with `run_setup` where the `conn_fun` argument 
 was not utilized.  
 * Added SQL script for post-processing of the MRHIER